% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_weighted_mds.R
\name{plot.mds}
\alias{plot.mds}
\title{Plot Weighted MDS}
\usage{
\method{plot}{mds}(mds, dimentions, group_by, ...)
}
\arguments{
\item{mds}{a list containing Xtilde and lambda and the weights}

\item{dimentions}{the dimentions to be plotted ex: c(1,2)}

\item{group_by}{a vector containing a group for each individual}

\item{...}{graphical params}
}
\description{
Plot Weighted MDS
}
